+++ src/app/containers/app-navbar/app-navbar-menu/app-navbar-menu.component.ts
@@ -7,6 +7,7 @@ import {
   EventEmitter,
   HostListener
 } from '@angular/core';
+import { expandFadeInAnimation } from '@shared/animations/fade-in.animation';
 
 enum Key {
   Backspace = 8,
@@ -22,16 +23,19 @@ enum Key {
 
 @Component({
   selector: 'app-navbar-menu',
+  animations: [expandFadeInAnimation],
   template: `
     <button class="btn btn-navbar btn-link ux-maker btn-toggle"
       (click)="toggleMenu()">
       <icon name="ellipsis-v"></icon>
       <icon *ngIf="appVersion.isNewAvailable" name="dot-circle-o" class="pulse update-indicator text-primary"></icon>
     </button>
+    <div class="menu-backdrop" *ngIf="!hide" (click)="hideMenu()"></div>
     <div class="panel menu-dropdown"
-      [class.slideInDown]="!hide"
+      [class.end-animation]="end"
+      [@expandFadeIn]="menuState"
+      (@expandFadeIn.done)="endAnimation($event)"
       >
-      <div class="menu-backdrop" *ngIf="!hide" (click)="hideMenu()"></div>
       <div class="list-group">
         <div *ngIf="appVersion.isNewAvailable" class="list-group-item">
           <button class="btn btn-success" title="click to update Echoes"
@@ -76,7 +80,12 @@ enum Key {
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AppNavbarMenuComponent implements OnInit {
+  end = false;
   hide = true;
+  get menuState() {
+    return this.hide ? 'hide' : 'show';
+  }
+
   @Input() signedIn = false;
   @Input()
   appVersion = {
@@ -110,6 +119,7 @@ export class AppNavbarMenuComponent implements OnInit {
   }
 
   toggleMenu() {
+    this.end = false;
     this.hide = !this.hide;
   }
 
@@ -123,5 +133,12 @@ export class AppNavbarMenuComponent implements OnInit {
 
   updateTheme(theme) {
     this.themeChange.emit(theme);
+  }
+
+  endAnimation({ toState }) {
+    if (toState === 'hide') {
+      this.end = true;
+    }
+    console.log('animation done!');
   }
 }