+++ src/app/containers/app-search/app-search.routing.ts
@@ -2,13 +2,14 @@
 import { ModuleWithProviders } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { YoutubeVideosComponent } from './youtube-videos.component';
+import { YoutubeVideosComponent } from './youtube-videos';
 import { AppSearchComponent } from './app-search.component';
-import { YoutubePlaylistsComponent } from './youtube-playlists.component';
+import { YoutubePlaylistsComponent } from './youtube-playlists';
 
 export const routing: ModuleWithProviders = RouterModule.forChild([
   {
-    path: 'search', component: AppSearchComponent,
+    path: 'search',
+    component: AppSearchComponent,
     children: [
       { path: '', redirectTo: 'videos', pathMatch: 'full' },
       { path: 'videos', component: YoutubeVideosComponent },